requirements.txt
==============================================================================
퀀트 자산 관리 시스템 v3.2 의존성 라이브러리 목록
==============================================================================
이 프로젝트를 실행하기 위해 필요한 모든 파이썬 라이브러리입니다.
아래 명령어를 사용하여 모든 라이브러리를 한 번에 설치할 수 있습니다:
pip install -r requirements.txt
------------------------------------------------------------------------------
--- 1. 코어 및 데이터 핸들링 (Core & Data Handling) ---
데이터 분석, 조작, 수치 계산의 근간이 되는 필수 라이브러리
pandas==2.2.2
numpy==1.26.4
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9  # PostgreSQL 데이터베이스 드라이버

--- 2. API 통신 및 환경 설정 (API & Environment) ---
외부 API 호출 및 .env 파일 관리를 위한 라이브러리
requests==2.32.3
python-dotenv==1.0.1

--- 3. 머신러닝 (Machine Learning) ---
특징 공학, 클러스터링, 예측 모델 훈련 및 평가를 위한 표준 라이브러리
scikit-learn==1.5.0
xgboost==2.0.3
lightgbm==4.3.0
joblib==1.4.2           # 훈련된 모델 저장 및 로드를 위함

--- 4. 딥러닝 - 시계열 예측 (Deep Learning for Time Series) ---
시계열 예측 모델 구현을 위한 핵심 프레임워크와 전문 라이브러리
torch==2.3.1            # PyTorch: 딥러닝 프레임워크 (CPU 버전 기준)
darts==0.28.0           # PyTorch 기반의 시계열 예측 전문 라이브러리

--- 5. 퀀트 분석 및 백테스팅 (Quantitative Analysis & Backtesting) ---
기술적 분석, 통계 모델, 백테스팅 및 성과 평가를 위한 라이브러리
pandas-ta==0.3.14b      # 기술적 지표 계산
statsmodels==0.14.2     # Fama-French 등 통계 회귀분석
backtrader==1.9.76.123  # 백테스팅 엔진
pyfolio-reloaded==0.2.3 # 백테스팅 결과 분석 및 시각화 (최신 유지보수 버전)

--- 6. 시각화 및 대시보드 (Visualization & Dashboard) ---
사용자 인터페이스 및 분석 결과 시각화를 위한 라이브러리
streamlit==1.35.0       # 메인 웹 대시보드 프레임워크
plotly==5.22.0          # 인터랙티브 차트
matplotlib==3.9.0       # 정적 차트 (pyfolio 의존성)
seaborn==0.13.2         # 통계 시각화